UI-Grid column size fix.
Although I created a method to dynamically resize the grid, the horixzontal scrollbar would not move from the original position. Several attempts were made to get this to work. Ill put this down a s a bug in the UI GRID sw. I stepped back from this, back to a grib of 10 rows where the grid is fixed in size, and hence the scrollbar is where it should be.