Register User begun, active field active to user entity
List of retailers added to login controller to provide drop down list on register user form.

User entity updated to include the attr 'active'. When a retail manager registers, active = false, the administrator must enable the account belore access is granted to the system. This creates a knock on requirement to check at login.
Will do this in REST, returning failed login attempt if the use is not active.