Rating : Average added to base of column.
My use of Globals did not allow for injecting uiGridConstants.
Instead the column was dynamically added as the grid was created in the ratings.controller.
The default behaviour for adding a column, placed it at the end, however the grid did not show this unless you panned across.
array.splice was used to insert into the middle of the array instead.